- Remove LapInstructure Project (in wrong file path anyway) - Create library of Test.Microsoft.Identity.LabInfrastructure (@bgavrilMS 's suggestion) - update namespaces and references
