+++ adal/tests/Test.ADAL.NET.UIAutomation/XamarinDroidADALTests.cs
@@ -2,7 +2,7 @@
 using System.IO;
 using System.Linq;
 using System.Threading;
-using LabInfrastructure;
+using Test.Microsoft.Identity.LabInfrastructure;
 using NUnit.Framework;
 using Test.Microsoft.Identity.Core.UIAutomation;
 using Xamarin.UITest;
