+++ core/tests/Test.Microsoft.Identity.Core.UIAutomation/XamarinUITestController.cs
@@ -25,7 +25,7 @@
 //
 //------------------------------------------------------------------------------
 
-using LabInfrastructure;
+using Test.Microsoft.Identity.LabInfrastructure;
 using NUnit.Framework;
 using System;
 using System.Linq;
