+++ core/tests/Test.Microsoft.Identity.LabInfrastructure/KeyVaultConfiguration.cs
@@ -25,18 +25,38 @@
 //
 //------------------------------------------------------------------------------
 
-using System.Collections.Generic;
-
-namespace LabInfrastructure
+namespace Test.Microsoft.Identity.LabInfrastructure
 {
-    public class UserQueryParameters
+    public class KeyVaultConfiguration
     {
-        public FederationProvider? FederationProvider { get; set; }
-        public bool? IsMamUser { get; set; }
-        public bool? IsMfaUser { get; set; }
-        public ISet<string> Licenses { get; set; }
-        public bool? IsFederatedUser { get; set; }
-        public UserType? IsUserType { get; set; }
-        public bool? IsExternalUser { get; set; }
+        /// <summary>
+        /// The URL of the Key Vault instance.
+        /// </summary>
+        public string Url { get; set; }
+
+        /// <summary>
+        /// The authentication type to use to communicate with the Key Vault.
+        /// </summary>
+        public KeyVaultAuthenticationType AuthType { get; set; }
+
+        /// <summary>
+        /// The ID of the test harness client.
+        /// </summary>
+        /// <remarks>
+        /// This should be configured as having access to the Key Vault instance specified at <see cref="Url"/>.
+        /// </remarks>
+        public string ClientId { get; set; }
+
+        /// <summary>
+        /// The thumbprint of the <see cref="System.Security.Cryptography.X509Certificates.X509Certificate2"/> to use when
+        /// <see cref="AuthType"/> is <see cref="KeyVaultAuthenticationType.ClientCertificate"/>.
+        /// </summary>
+        public string CertThumbprint { get; set; }
+    }
+
+    public enum KeyVaultAuthenticationType
+    {
+        ClientCertificate,
+        UserCredential
     }
 }
