+++ msal/tests/Test.MSAL.NET.UIAutomation/Test.MSAL.NET.UIAutomation.csproj
@@ -42,9 +42,9 @@
       <Project>{3DC6EC76-D350-4D43-B206-A4CEC8AA36D4}</Project>
       <Name>Test.Microsoft.Identity.Core.UIAutomation</Name>
     </ProjectReference>
-    <ProjectReference Include="..\..\..\LabInfrastructure\LabInfrastructure.csproj">
-      <Project>{8af6b089-8aea-41c9-8890-c94b3e8876a0}</Project>
-      <Name>LabInfrastructure</Name>
+    <ProjectReference Include="..\..\..\core\tests\Test.Microsoft.Identity.LabInfrastructure\Test.Microsoft.Identity.LabInfrastructure.csproj">
+      <Project>{a181778d-5917-41ce-aa5f-7daaa7b7f5bb}</Project>
+      <Name>Test.Microsoft.Identity.LabInfrastructure</Name>
     </ProjectReference>
     <ProjectReference Include="..\dev apps\XForms\XForms\XForms.csproj">
       <Project>{8d14a8ff-6cb9-44b1-b7c4-3c1d79ce8412}</Project>
